requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
python-dotenv==1.0.1
jinja2==3.1.4
pydantic>=2.10.3
Pillow==10.4.0
